+++ doc/tutorials/space_shooter/app/assets/images/bullet.png
