+++ doc/tutorials/space_shooter/app/assets/images/enemy.png
