+++ doc/tutorials/space_shooter/app/assets/images/explosion.png
