+++ doc/tutorials/space_shooter/app/assets/images/player.png
