+++ doc/tutorials/space_shooter/app/assets/images/stars.png
