+++ doc/tutorials/space_shooter/app/lib/step_1/main.dart
@@ -2,6 +2,10 @@ import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
 
+void main() {
+  runApp(GameWidget(game: SpaceShooterGame()));
+}
+
 class Player extends PositionComponent {
   static final _paint = Paint()..color = Colors.white;
   @override
@@ -20,14 +24,5 @@ class SpaceShooterGame extends FlameGame {
         ..height = 100
         ..anchor = Anchor.center,
     );
-  }
-}
-
-class MyGame extends StatelessWidget {
-  const MyGame({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return GameWidget(game: SpaceShooterGame());
   }
 }
