+++ doc/tutorials/space_shooter/app/lib/step_2/main.dart
@@ -1,0 +1,44 @@
+import 'package:flame/components.dart';
+import 'package:flame/game.dart';
+import 'package:flame/input.dart';
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(GameWidget(game: SpaceShooterGame()));
+}
+
+class SpaceShooterGame extends FlameGame with PanDetector {
+  late Player player;
+
+  @override
+  Future<void> onLoad() async {
+    await super.onLoad();
+
+    player = Player();
+
+    add(player);
+  }
+
+  @override
+  void onPanUpdate(DragUpdateInfo info) {
+    player.move(info.delta.game);
+  }
+}
+
+class Player extends SpriteComponent with HasGameRef<SpaceShooterGame> {
+  @override
+  Future<void> onLoad() async {
+    await super.onLoad();
+
+    sprite = await gameRef.loadSprite('player-sprite.png');
+
+    position = gameRef.size / 2;
+    width = 100;
+    height = 150;
+    anchor = Anchor.center;
+  }
+
+  void move(Vector2 delta) {
+    position.add(delta);
+  }
+}
