+++ doc/tutorials/space_shooter/app/pubspec.yaml
@@ -6,22 +6,17 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: '>=2.18.2 <3.0.0'
 
 dependencies:
-  dashbook: ^0.1.5
   flame: ^1.4.0
   flutter:
     sdk: flutter
-  flutter_highlight: ^0.7.0
-  flutter_markdown: ^0.6.7
-  markdown: ^6.0.0
 
 dev_dependencies:
   flame_lint: ^0.1.3
-  flutter_lints: ^1.0.0
   flutter_test:
-    sdk: flutter
+    sdk: flutter 
 
 flutter:
   uses-material-design: true
