+++ doc/tutorials/space_shooter/app/web/index.html
