+++ doc/tutorials/space_shooter/space_shooter.md
@@ -1,0 +1,17 @@
+# Space Shooter Game Tutorial
+
+In this tutorial, we will follow a step-by-step process for coding a game using the Flame
+engine.
+
+This tutorial assumes that you have at least some familiarity with common programming concepts, and
+with the [Dart] programming language.
+
+
+[Dart]: https://dart.dev/overview
+
+```{toctree}
+:hidden:
+
+1. Getting Started  <step_1.md>
+2. Controlling the player and adding some graphics <step_2.md>
+```
