--- tutorials/space_shooter/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: ffb2ecea5223acdd139a5039be2f9c796962833d
-  channel: stable
-
-project_type: app
