--- tutorials/space_shooter/README.md
@@ -1,3 +1,0 @@
-# Flame tutorial
-
-In this Flame tutorial we build a space shooter game which covers several parts of the engine.
