--- tutorials/space_shooter/lib/steps/1_getting_started/1_flame_game/step.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:tutorials_space_shooter/steps/1_getting_started/1_flame_game/code.dart';
-import 'package:tutorials_space_shooter/steps/1_getting_started/1_flame_game/tutorial.dart';
-import 'package:tutorials_space_shooter/widgets/step_scaffold.dart';
-
-class RunningFlameStep extends StatelessWidget {
-  const RunningFlameStep({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const StepScaffold(
-      tutorial: tutorial,
-      game: MyGame(),
-    );
-  }
-}
