--- tutorials/space_shooter/lib/steps/1_getting_started/2_input_and_graphics/code.dart
@@ -1,45 +1,0 @@
-import 'package:flame/components.dart';
-import 'package:flame/game.dart';
-import 'package:flame/input.dart';
-import 'package:flutter/material.dart';
-
-class Player extends SpriteComponent with HasGameRef<SpaceShooterGame> {
-  @override
-  Future<void> onLoad() async {
-    sprite = await gameRef.loadSprite('player-sprite.png');
-
-    position = gameRef.size / 2;
-    width = 100;
-    height = 150;
-    anchor = Anchor.center;
-  }
-
-  void move(Vector2 delta) {
-    position.add(delta);
-  }
-}
-
-class SpaceShooterGame extends FlameGame with PanDetector {
-  late Player player;
-
-  @override
-  Future<void>? onLoad() {
-    player = Player();
-    add(player);
-    return null;
-  }
-
-  @override
-  void onPanUpdate(DragUpdateInfo info) {
-    player.move(info.delta.game);
-  }
-}
-
-class MyGame extends StatelessWidget {
-  const MyGame({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return GameWidget(game: SpaceShooterGame());
-  }
-}
